Fixed problem with shadows (16 bit -> 32 bit for gbuffer)

+++ src/Window/Window.cpp
@@ -204,7 +204,7 @@ void Window::startLoop()
     // буфер позиций
     glGenTextures(1, &gPosition);
     glBindTexture(GL_TEXTURE_2D, gPosition);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, Window::_width, Window::_height, 0, GL_RGB, GL_FLOAT, NULL);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, Window::_width, Window::_height, 0, GL_RGB, GL_FLOAT, NULL);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, gPosition, 0);
@@ -212,7 +212,7 @@ void Window::startLoop()
     // буфер нормалей
     glGenTextures(1, &gNormal);
     glBindTexture(GL_TEXTURE_2D, gNormal);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, Window::_width, Window::_height, 0, GL_RGB, GL_FLOAT, NULL);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, Window::_width, Window::_height, 0, GL_RGB, GL_FLOAT, NULL);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT1, GL_TEXTURE_2D, gNormal, 0);
@@ -268,9 +268,7 @@ void Window::startLoop()
 
     ShadowMap shadowMap(state, 2048, 2048, 25);
     LocalPlayer player(glm::vec3(0, 50, 0));
-    //NetPlayer netPlayer(glm::vec3(2, 6, 2), state);
-    //
-    //PlayerRenderer netPlayerRenderer(&netPlayer);
+
 
     std::ifstream cfgFile("settings.cfg");
     std::string host = "127.0.0.1";
